Fixed overflowing

+++ EncryptionApp/Extensions.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Numerics;
 
 namespace System
 {
@@ -13,9 +14,9 @@ namespace System
         /// </summary>
         /// <param name="n"></param>
         /// <returns></returns>
-        public static bool IsPrime(this int n)
+        public static bool IsPrime(this BigInteger n)
         {
-            int i = 2;
+            BigInteger i = 2;
             while(i*i <= n)
             {
                 if(n % i == 0)
@@ -33,7 +34,7 @@ namespace System
         /// <param name="a"></param>
         /// <param name="b"></param>
         /// <returns></returns>
-        public static int GreatestCommonDivisor(this int a, int b)
+        public static BigInteger GreatestCommonDivisor(this BigInteger a, BigInteger b)
         {
             return gcd(a, b);
         }
@@ -44,7 +45,7 @@ namespace System
         /// <param name="a"></param>
         /// <param name="b"></param>
         /// <returns></returns>
-        public static int SolveDiophantineEquation(this int a, int b, out int x, out int y)
+        public static BigInteger SolveDiophantineEquation(this BigInteger a, BigInteger b, out BigInteger x, out BigInteger y)
         {
             return egcd(a, b, out x, out y);
         }
@@ -56,9 +57,9 @@ namespace System
         /// <param name="power"></param>
         /// <param name="module"></param>
         /// <returns></returns>
-        public static int BinaryPow(this int a, int power, int module)
+        public static BigInteger BinaryPow(this BigInteger a, BigInteger power, BigInteger module)
         {
-            int ans = 1;
+            BigInteger ans = 1;
             while(power > 0)
             {
                 if(power % 2 == 1)
@@ -74,21 +75,21 @@ namespace System
         }
 
         // алгоритм Евклида
-        private static int gcd(int a, int b)
+        private static BigInteger gcd(BigInteger a, BigInteger b)
         {
             return b == 0 ? a : gcd(b, a % b);
         }
 
         // расширенный алгоритм Евклида, решает уравнение a*x + b*y = gcd(a, b);
-        private static int egcd(int a, int b, out int x, out int y)
+        private static BigInteger egcd(BigInteger a, BigInteger b, out BigInteger x, out BigInteger y)
         {
             if (a == 0)
             {
                 x = 0; y = 1;
                 return b;
             }
-            int x1, y1;
-            int d = egcd(b % a, a, out x1, out y1);
+            BigInteger x1, y1;
+            BigInteger d = egcd(b % a, a, out x1, out y1);
             x = y1 - (b / a) * x1;
 
             y = x1;
